Rename 'components/WhatsNew.vue' to 'components/index/SiteTopUpper/WhatsNewCard.vue'
