+++ components/index/SiteTopUpper/WhatsNewCard.vue
